biplotEZ: EZ-to-Use Biplots

Details

The goal of biplotEZ is to provide users an EZ-to-use platform for visually representing their data with biplots. Currently, this package includes principal component analysis (PCA) and canonical variate analysis (CVA) biplots. This is accompanied by various formatting options for the samples and axes. Alpha-bags and concentration ellipses are included for visual enhancements and interpretation.

Details

Package:	biplotEZ

Type:	Package

Version:	2.0

Date:	05-04-2024

License:	MIT

LazyLoad:	TRUE

Author(s)

Sugnet Lubbe (Maintainer, muvisu@sun.ac.za)

Niël le Roux

Johané Nienkemper-Swanepoel

Raeesa Ganey

Ruan Buys

Zoë-Mae Adams

Peter Manefeldt

Core Functions

biplot

PCA

CVA

CA

Code Availability

The newest version of the package can be obtained on GitHub: https://github.com/MuViSU/biplotEZ